+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- New config: `cli` to disable a few rules for files in `bin`, `scripts` and `cli` directories, enabled by default.
 - `jest`/`vitest`: by default include nested `__test(s)__` directories and `[-_].spec.*` files.
 - `sonar`: change `prefer-single-boolean-return` default severity to `warn`.